fix for big size image

+++ libEM/emdata_transform.cpp
@@ -1791,8 +1791,8 @@ void EMData::insert_clip(const EMData * const block, const IntPoint &origin) {
 	if (x1 < x0 || y1 < y0 || z1 < z0) return; // out of bounds, this is fine, nothing happens
 
 	size_t clipped_row_size = (x1-x0) * sizeof(float);
-	int src_secsize =  nx1 * ny1;
-	int dst_secsize = nx * ny;
+	size_t src_secsize =  (size_t)(nx1 * ny1);
+	size_t dst_secsize = (size_t)(nx * ny);
 
 /*
 #ifdef EMAN2_USING_CUDA
@@ -1818,8 +1818,8 @@ void EMData::insert_clip(const EMData * const block, const IntPoint &origin) {
 	float *src = block->get_data() + zd0 * src_secsize + yd0 * nx1 + xd0;
 	float *dst = get_data() + z0 * dst_secsize + y0 * nx + x0;
 	
-	int src_gap = src_secsize - (y1-y0) * nx1;
-	int dst_gap = dst_secsize - (y1-y0) * nx;
+	size_t src_gap = src_secsize - (y1-y0) * nx1;
+	size_t dst_gap = dst_secsize - (y1-y0) * nx;
 	
 	for (int i = z0; i < z1; i++) {
 		for (int j = y0; j < y1; j++) {
